requirements.txt
psycopg2
mysql-connector-python
scrapy
boto3
pymongo
scrapy-user-agents